refactor(web): extract duplicate createRange helper into shared utils
