+++ apps/web/src/intellisense/utils.ts
@@ -1,0 +1,16 @@
+import type * as Monaco from "monaco-editor";
+
+export function createRange(
+  monaco: typeof Monaco,
+  model: Monaco.editor.ITextModel,
+  position: Monaco.Position,
+) {
+  const word = model.getWordUntilPosition(position);
+
+  return new monaco.Range(
+    position.lineNumber,
+    word.startColumn,
+    position.lineNumber,
+    word.endColumn,
+  );
+}
